api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
tensorflow-cpu==2.16.1
pillow==10.4.0
numpy==1.26.4
requests==2.32.3
python-dotenv==1.0.1
python-multipart==0.0.9
firebase-admin==6.5.0
google-cloud-storage==2.17.0
google-cloud-firestore==2.16.0